@@ -1,30 +1,27 @@
 {
-  "name": "svelte-infinite-scroll-selector",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "build": "rollup -c",
-    "dev": "rollup -c -w",
-    "start": "sirv public --no-clear",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
-  },
-  "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.0.0",
-    "@rollup/plugin-node-resolve": "^11.0.0",
-    "rollup": "^2.3.4",
-    "rollup-plugin-css-only": "^3.1.0",
-    "rollup-plugin-livereload": "^2.0.0",
-    "rollup-plugin-svelte": "^7.0.0",
-    "rollup-plugin-terser": "^7.0.0",
-    "svelte": "^3.0.0",
-    "svelte-check": "^2.0.0",
-    "svelte-preprocess": "^4.0.0",
-    "@rollup/plugin-typescript": "^8.0.0",
-    "typescript": "^4.0.0",
-    "tslib": "^2.0.0",
-    "@tsconfig/svelte": "^2.0.0"
-  },
-  "dependencies": {
-    "sirv-cli": "^2.0.0"
-  }
+	"name": "svelte-infinite-scroll-selector",
+	"version": "1.0.0",
+	"description": "",
+	"main": "index.js",
+	"scripts": {
+		"build": "rollup -c",
+		"dev": "rollup -c -w"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/jobayer977/svelte-infinite-scroll-selector.git"
+	},
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
+	"bugs": {
+		"url": "https://github.com/jobayer977/svelte-infinite-scroll-selector/issues"
+	},
+	"homepage": "https://github.com/jobayer977/svelte-infinite-scroll-selector#readme",
+	"devDependencies": {
+		"rollup": "^2.76.0",
+		"rollup-plugin-node-resolve": "^5.2.0",
+		"rollup-plugin-svelte": "^7.1.0",
+		"svelte": "^3.49.0"
+	}
 }
